requirements.txt
fastapi
uvicorn
streamlit
google-generativeai
python-dotenv
pydantic
requests
streamlit-option-menu